requirements.txt
git+https://github.com/NimaBoscarino/AOT-GAN-for-Inpainting.git
streamlit_drawable_canvas
opencv-python
torch
torchvision
pillow
transformers
